Fix GitHub actions and extend pipeline
Installs required packages in runner to make build successful.
Extends the pipeline to run on Release and Debug build flavors and
use a cache to improve pipeline TAT.
This includes running a few unit tests only in Debug mode since they
rely on overflow panics and adding a test that also covers the case
of a too large map in Release builds.