Implement framework for general spawn lists and room-based map modifiers
In room-based maps, player starting position and other entities may be
(randomly) determined on a per-room basis.
Introduce treasure chests as immovable spawn types. The goal is to have
only a single treasure chest per level as the goal for the player to reach.
For now, there are chests everywhere and they may also be rendered on top
of the player.