Introduce an 'objective spawner' to spawn a single known entity in a level
Use this objective spawner to place a single treasure chest per level.
Also refactor spawning code a bit and fix an issue where items could be
rendered on top of the player by introducing fixed 'Z-buffer levels' for
different entities and fixing the spawning code to use correct initial
translation for the sprites.