Allow the player character to push monsters out of the way
This serves as a rudimentary form of conflict resolution instead of
traditional combat to allow the player to move past enemies.
The player can push monsters around Sokoban style - but the monsters
will also push back occasionally. Due to the random monster movement
and the essentially simultaneous turn execution, this is not the
funnest implementation, yet.