Try using sprites for UI elements
Attempt at assemblin ga sizable / fillable bar, e.g. for hit points or remaining
turns, from multiple sprites. Unfortunately, this is not straightforward in egui
because it does not allow for straightforward absolute positioning.
Using native Bevy UI widgets may be more suitable for this but requires lots of
boilerplate.